pyproject.toml
[tool.poetry]
name = "fresh-cart"
version = "0.1.0"
description = "Shop like you never did before"
authors = ["SCADEMY <dev@scademy.com>"]
license = "Proprietary"

[tool.poetry.dependencies]
python = "3.11.3"
Flask = "2.2.3"
alembic = "1.10.3"
psycopg2 = "2.9.6"
openai = "0.27.4"
Flask-SocketIO = "5.3.3"
Flask-SQLAlchemy = "3.0.3"
pgvector = "0.1.6"
tiktoken = "0.3.3"
Flask-Cors = "3.0.10"
Flask-JWT-Extended = "4.4.4"

[tool.poetry.dev-dependencies]
black = "23.3.0"
isort = "5.12.0"
autoflake = "2.1.1"
nose2 = "0.12.0"
pytruth = "1.1.0"

[tool.black]
line-length = 79

[tool.isort]
multi_line_output = 3
include_trailing_comma = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
